Convert SketchColorPickerController (#4128)
* converted

* lint

* update tests

* snapshot for bumped button

* fix sizeForItemAt not called issue

* fix snapshot

* fix for comments

+++ Wire-iOS/Sources/UserInterface/Sketchpad/CanvasViewController.swift
@@ -174,7 +174,7 @@ final class CanvasViewController: UIViewController, UINavigationControllerDelega
         colorPickerController.willMove(toParent: self)
         view.addSubview(colorPickerController.view)
         addChild(colorPickerController)
-        colorPickerController.selectedColorIndex = UInt(colorPickerController.sketchColors.firstIndex(of: UIColor.accent()) ?? 0)
+        colorPickerController.selectedColorIndex = colorPickerController.sketchColors.firstIndex(of: UIColor.accent()) ?? 0
     }
     
     func createConstraints() {
